Add module for operating on cartesian coordinates
I wrote the code originally as helpers for rendering Command Card
arrows. But it become large enough to deserve its own module. On top of
that, these will likely come to be used by many other components, like
units and terrain.

Another small change was creating a single function for generating the
arrow paths. This gives me just one entry point to edit should I need to
refactor later. It also gives me the convenience of calling helpers at
the top of the template file. Don't know if I'll stick with that
convention, but I'll try it for now.