Add vector module and Point.from_polar func
Vectors are needed for calculating new location of a point as it moves
along a straight line or along an arc. The immediate need is for each of
creating the Command Card arrow svgs, but it will also see a lot of use
when calculating the positions of units in the world.