Add y-intercept function to vector module
This helps with drawing the svg for the command card arrows. The
quadratic curve control point will lie on the y-axis inline with the
arrow end vector. This is before the whole arrow gets translated into
the center of the svg (It starts from the 0,0 point and continues off
the canvas initially).