Add func to displace point(s); Fix y-intercept tests
Displacement will be useful for mapping e.g. the elements of a unit to a
new location. This will allow me to keep a clean master version of, say,
where a ship's turrets are relative to the ship's center; then I can
displace all those points to the ship's current location.