Fix ship's path and center of rotation
The ship now enters and exits offscreen and passes between the islands,
like a good boat should. I also moved extended the ship's hull's svg in
order to put the center of the viewbox (and hence, the center of
rotation) at approx 25% behind the ship's bow. More realistic that way.