Implement Slider sync
This is very easy to do since I was already tracking the actual elapsed
seconds. I just had to create a new dedicated timeline for the slider
and pass the actual time into the seek method. Easy peasy.

I think this whole LiveView/GSAP stack is going to work very well
indeed. So easy to add animations to a timeline as soon as an element is
mounted to the DOM. It'll be just as easy to remove animations from the
timeline if an element gets removed by LiveView as well. Very
declarative. I love it.