Add Arena template
This template renders the game area 1:1. It will also show the outline
of the battle area. It's also the parent element of all terrain and
units.

Decks and other UI will be handled in another template.

Really, this template could have been called 'world', but that's already
taken for now. That may change in the future...