Add a Viewbox module
I'm starting to think I need a little different organization of all this
geometry stuff. I think a Path module is needed. A Path would shoulder a
lot of the burden that Command currently carries. It would house
functions for calculating the viewbox for example.

I also have to refactor the Point and Vector modules. Points as tuples
is turning out to be really annoying. Vector also probably doesn't need
the nesting involved in having the point be yet another map. Keep things
flat.