Change all SVG numbers to integers
The primary reason for this is ease of testing. I was tired of not
knowing whether a number would turn out to be a float or an integer.

Later, if I deem it necessary, I might change my approach from
all-integer to all-single-decimal-float.

I think the map is large enough where fractions of a pixel won't be
noticeable.