Convert commands to segments
This was a big one. There is a good chunk of the infrastructure here now
to be able to later painlessly convert issued commands into segments
that are easily consumed by the liveview.

Things are a bit messy right now. The next several commint will likely
be refactors as I clean up after myself.