Add CommandIds struct (and many other getters)
I figure the ids of unit/card/segment will be passed around the app a
bunch. If so, it'll be helpful to have this simple struct around for
pattern matching, etc. Really, this struct fully describes every issued
command in the game. It encodes all the information needed for looking
up all the command's details.

This commit also contains many getters I discovered I needed while
writing the Mission test.