Delete card and deck
The Command module is going to be front and center from now on. The
Command struct will be a token. This way I can keep one master list per
unit and not worry about moving commands from deck.hand to deck.discard
and then copying it to a command queue. Instead, to play a cards, just
modify it and keep it in the same list. Then iterate and filter over
that list to come up with current hand and discards, etc.