Get panning working again
I still have to get capturing working again. Right now, if you're
panning and quickly pass over e.g. a button, the move is cancelled. I
had this fixed in an earlier commit, but removed it b/c it was
interfering with the clicking of command squares in Chromium browsers.